+++ app/src/main/java/com/example/naebaecamteam1rm_spartube/homepage/HomeAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.naebaecamteam1rm_spartube.search
+package com.example.naebaecamteam1rm_spartube.homepage
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -6,24 +6,23 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+
 import com.example.naebaecamteam1rm_spartube.data.TubeDataModel
 import com.example.naebaecamteam1rm_spartube.databinding.ItemRecyclerviewBinding
 
-class SearchAdapter(private val context: Context) : RecyclerView.Adapter<SearchAdapter.Holder>() {
-    var items = ArrayList<TubeDataModel>()
+class HomeAdapter(context: Context) : RecyclerView.Adapter<HomeAdapter.Holder>() {
+
+    var list = ArrayList<TubeDataModel>()
+
     var mContext = context
 
-
-    fun clearItem() {
-        items.clear()
-        notifyDataSetChanged()
-    }
-
     interface ItemClick {
-        fun onClick(view : View, position : Int)
+
+        fun onClick(view : View, tubeData : TubeDataModel)
+
     }
 
-    var itemClick : ItemClick? = null
+    var itemClick: ItemClick? = null
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
         return Holder(
@@ -34,26 +33,39 @@ class SearchAdapter(private val context: Context) : RecyclerView.Adapter<SearchA
     }
 
     override fun onBindViewHolder(holder: Holder, position: Int) {
-        var currentItem = items[position]
-        holder.bind(currentItem)
+        var item = list[position]
+        holder.bind(item)
+
+//        holder.itemView.setOnClickListener {  //클릭이벤트추가부분
+//            itemClick?.onClick(it, position)
+//        }
+//        holder.thumbnail.setImageResource(list[position].url)
+//        holder.title.text = list[position].title
     }
 
     override fun getItemId(position: Int): Long {
         return position.toLong()
     }
 
-    override fun getItemCount() = items.size
+    override fun getItemCount(): Int {
+        return list.size
+    }
 
-    inner class Holder(val binding: ItemRecyclerviewBinding) : RecyclerView.ViewHolder(binding.root) {
+    inner class Holder(val binding: ItemRecyclerviewBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+        //        val thumbnail = binding.ivThumbnails
+        //        val title = binding.tvTitle
 
         fun bind(item: TubeDataModel) = with(binding){ //클릭이벤트추가부분
             itemView.setOnClickListener{
-                itemClick?.onClick(it, adapterPosition)
+                itemClick?.onClick(it, item)
             }
             Glide.with(mContext)
                 .load(item.thumbnail)
                 .into(ivThumbnails)
             tvTitle.text = item.title
+
+
         }
     }
 }
