+++ app/src/main/java/com/example/naebaecamteam1rm_spartube/main/MainActivity.kt
@@ -53,5 +53,10 @@ class MainActivity : AppCompatActivity() {
         myPageFragment.addItem(item)
     }
 
+//    fun addPlayList(item: TubeDataModel?){
+//        val playListFragment = viewPagerAdapter.getPlayListFragment()
+//        playListFragment.addItem(item)
+//    }
+
 
 }
