+++ app/src/main/java/com/example/naebaecamteam1rm_spartube/playlistpage/PlayListAdapter.kt
@@ -1,0 +1,4 @@
+package com.example.naebaecamteam1rm_spartube.playlistpage
+
+class PlayListAdapter {
+}
