+++ app/src/main/java/com/example/naebaecamteam1rm_spartube/playlistpage/PlayListFragment.kt
@@ -1,14 +1,18 @@
-package com.example.naebaecamteam1rm_spartube
+package com.example.naebaecamteam1rm_spartube.playlistpage
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.example.naebaecamteam1rm_spartube.R
+import com.example.naebaecamteam1rm_spartube.mypage.MyPageViewModel
 
 
+class PlayListFragment : Fragment() {
 
-class PlaylistFragment : Fragment() {
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -16,4 +20,10 @@ class PlaylistFragment : Fragment() {
     ): View? {
         return inflater.inflate(R.layout.fragment_playlist, container, false)
     }
+
+//    fun addItem(item: TubeDataModel?) {
+//
+//    }
+
+
 }
