+++ app/src/main/java/com/example/naebaecamteam1rm_spartube/searchpage/SearchAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.naebaecamteam1rm_spartube.home
+package com.example.naebaecamteam1rm_spartube.searchpage
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -6,23 +6,24 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-
 import com.example.naebaecamteam1rm_spartube.data.TubeDataModel
 import com.example.naebaecamteam1rm_spartube.databinding.ItemRecyclerviewBinding
 
-class HomeAdapter(context: Context) : RecyclerView.Adapter<HomeAdapter.Holder>() {
-
-    var list = ArrayList<TubeDataModel>()
-
+class SearchAdapter(private val context: Context) : RecyclerView.Adapter<SearchAdapter.Holder>() {
+    var items = ArrayList<TubeDataModel>()
     var mContext = context
 
-    interface ItemClick {
 
-        fun onClick(view : View, tubeData : TubeDataModel)
-
+    fun clearItem() {
+        items.clear()
+        notifyDataSetChanged()
     }
 
-    var itemClick: ItemClick? = null
+    interface ItemClick {
+        fun onClick(view : View, tubeData : TubeDataModel)
+    }
+
+    var itemClick : ItemClick? = null
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
         return Holder(
@@ -33,28 +34,17 @@ class HomeAdapter(context: Context) : RecyclerView.Adapter<HomeAdapter.Holder>()
     }
 
     override fun onBindViewHolder(holder: Holder, position: Int) {
-        var item = list[position]
-        holder.bind(item)
-
-//        holder.itemView.setOnClickListener {  //클릭이벤트추가부분
-//            itemClick?.onClick(it, position)
-//        }
-//        holder.thumbnails.setImageResource(list[position].url)
-//        holder.title.text = list[position].title
+        var currentItem = items[position]
+        holder.bind(currentItem)
     }
 
     override fun getItemId(position: Int): Long {
         return position.toLong()
     }
 
-    override fun getItemCount(): Int {
-        return list.size
-    }
+    override fun getItemCount() = items.size
 
-    inner class Holder(val binding: ItemRecyclerviewBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        //        val thumbnails = binding.ivThumbnails
-        //        val title = binding.tvTitle
+    inner class Holder(val binding: ItemRecyclerviewBinding) : RecyclerView.ViewHolder(binding.root) {
 
         fun bind(item: TubeDataModel) = with(binding){ //클릭이벤트추가부분
             itemView.setOnClickListener{
@@ -64,8 +54,6 @@ class HomeAdapter(context: Context) : RecyclerView.Adapter<HomeAdapter.Holder>()
                 .load(item.thumbnail)
                 .into(ivThumbnails)
             tvTitle.text = item.title
-
-
         }
     }
 }
