+++ app/src/main/res/drawable/video_detail_page_btn_ic_like.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:drawable="@drawable/video_detail_page_ic_like"
+        android:width="20dp"
+        android:height="20dp" />
+</layer-list>
