+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".home.HomeFragment">
+    tools:context=".homepage.HomeFragment">
     <RelativeLayout
         android:id="@+id/relative_appbar"
         android:layout_width="match_parent"
