+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="left_padding">8dp</dimen>
+    <dimen name="top_padding">8dp</dimen>
+    <dimen name="right_padding">8dp</dimen>
+    <dimen name="bottom_padding">8dp</dimen>
+</resources>
